requirements.txt
vsg==3.12.0
vunit_hdl==4.7.0
pylint==3.3.2
flake8==7.1.1
